# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/startups": {
      "post": {
        "operationId": "StartupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "get": {
        "operationId": "StartupController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 20,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "analyzing",
                "pending_review",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "industry",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pre_seed",
                "seed",
                "series_a",
                "series_b",
                "series_c",
                "series_d",
                "series_e",
                "series_f_plus"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}": {
      "get": {
        "operationId": "StartupController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "patch": {
        "operationId": "StartupController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStartupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "delete": {
        "operationId": "StartupController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/submit": {
      "post": {
        "operationId": "StartupController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/resubmit": {
      "post": {
        "operationId": "StartupController_resubmit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/jobs": {
      "get": {
        "operationId": "StartupController_getJobs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/upload-url": {
      "post": {
        "operationId": "StartupController_getUploadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/draft": {
      "post": {
        "operationId": "StartupController_saveDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "get": {
        "operationId": "StartupController_getDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/progress": {
      "get": {
        "operationId": "StartupController_getProgress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/memo.pdf": {
      "get": {
        "operationId": "StartupController_downloadMemo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/report.pdf": {
      "get": {
        "operationId": "StartupController_downloadReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/data-room": {
      "post": {
        "operationId": "StartupController_uploadToDataRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDataRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "get": {
        "operationId": "StartupController_getDataRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/data-room/{docId}/permissions": {
      "patch": {
        "operationId": "StartupController_updateDataRoomPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataRoomPermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/data-room/{docId}": {
      "delete": {
        "operationId": "StartupController_deleteDataRoomDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/interest": {
      "get": {
        "operationId": "StartupController_getInvestorInterest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/interest/{interestId}/respond": {
      "post": {
        "operationId": "StartupController_respondToInterest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondInterestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/{id}/meetings": {
      "post": {
        "operationId": "StartupController_scheduleMeeting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleMeetingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      },
      "get": {
        "operationId": "StartupController_getMeetings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/approved/list": {
      "get": {
        "operationId": "StartupController_findApproved",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 20,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "industry",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pre_seed",
                "seed",
                "series_a",
                "series_b",
                "series_c",
                "series_d",
                "series_e",
                "series_f_plus"
              ]
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/approved/{id}": {
      "get": {
        "operationId": "StartupController_findApprovedById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/approved/{id}/evaluation": {
      "get": {
        "operationId": "StartupController_getEvaluation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/admin/all": {
      "get": {
        "operationId": "StartupController_adminFindAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 20,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "analyzing",
                "pending_review",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "industry",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pre_seed",
                "seed",
                "series_a",
                "series_b",
                "series_c",
                "series_d",
                "series_e",
                "series_f_plus"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/admin/pending": {
      "get": {
        "operationId": "StartupController_adminFindPending",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 20,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "analyzing",
                "pending_review",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "industry",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pre_seed",
                "seed",
                "series_a",
                "series_b",
                "series_c",
                "series_d",
                "series_e",
                "series_f_plus"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/admin/{id}/approve": {
      "post": {
        "operationId": "StartupController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/admin/{id}/reject": {
      "post": {
        "operationId": "StartupController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
    "/startups/admin/{id}/reanalyze": {
      "post": {
        "operationId": "StartupController_reanalyze",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Startup"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS Boilerplate API",
    "description": "Production-ready NestJS boilerplate with auth, queues, storage, and email",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "SubmitStartupDto": {
        "type": "object",
        "properties": {}
      },
      "PresignedUrlDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "fileType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "fileSize": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 52428800,
            "minimum": 0
          }
        },
        "required": [
          "fileName",
          "fileType",
          "fileSize"
        ]
      },
      "SaveDraftDto": {
        "type": "object",
        "properties": {
          "draftData": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "draftData"
        ]
      },
      "UploadDataRoomDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "assetId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "category"
        ]
      },
...
    }
  }
}